add services

+++ src/main/java/com/example/eshop/data/repositories/OrderItemRepository.java
@@ -3,5 +3,5 @@ package com.example.eshop.data.repositories;
 import com.example.eshop.entities.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
